// src/book/dtos/example-dto.ts
export class exampleDTO {
    id: number;
    name: string;
    description: string;
    price: number;
    publisherName: string; // Publisher's name
    writer: string;
    pageNumber: number;
    datePublished: number;
    language: string;
    bookDimension: string;
    barcode: string;
    isbn: string;
    editionNumber: string;
    imagePath: string;
    isDeleted: boolean;
}